class User change int id on Long id
interface UserDAO
change name UserDao

class UserDAOImpl
change name UserDaoImpl
change user1 on newUser
change show() delete detach(user)
delete @Transactional

class UserServiceImpl
add @Transactional

package javaConfig
change name config

+++ src/main/java/com/taskcomp/controllers/UserController.java
@@ -36,20 +36,20 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable("id") int id) {
+    public String delete(@PathVariable("id") Long id) {
         userService.delete(id);
         return "redirect:/users";
     }
 
     @GetMapping("/{id}/edit")
-    public String edit(Model model, @PathVariable("id") int id) {
+    public String edit(Model model, @PathVariable("id") Long id) {
         model.addAttribute("user", userService.show(id));
         return "edit";
     }
 
 
     @PatchMapping("/{id}")
-    public String update(@ModelAttribute("user") User user, @PathVariable("id") int id) {
+    public String update(@ModelAttribute("user") User user, @PathVariable("id") Long id) {
         userService.update(id, user);
         return "redirect:/users";
     }
